+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 [full changelog](http://github.com/yolk/aua/compare/v0.1.1...master)
 
+* Added support for unknown simple user agent string (e.g. my_app/1.0)
+
 ### 0.1.1 / 2011-01-28
 
 [full changelog](http://github.com/yolk/aua/compare/v0.1.0...v0.1.1)